// frontend/src/api/generated/models/return-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 返品種別
 * @export
 * @enum {string}
 */

export const ReturnType = {
    Inbound: 'INBOUND',
    Inventory: 'INVENTORY',
    Outbound: 'OUTBOUND'
} as const;

export type ReturnType = typeof ReturnType[keyof typeof ReturnType];