// frontend/src/api/generated/models/return-reason.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 返品理由コード
 * @export
 * @enum {string}
 */

export const ReturnReason = {
    QualityDefect: 'QUALITY_DEFECT',
    ExcessQuantity: 'EXCESS_QUANTITY',
    WrongDelivery: 'WRONG_DELIVERY',
    Expired: 'EXPIRED',
    Damaged: 'DAMAGED',
    Other: 'OTHER'
} as const;

export type ReturnReason = typeof ReturnReason[keyof typeof ReturnReason];